Package step improvements
* Use the electron-builder method of packaging
* Use `./dist/electron/main` as main file.
* Rewrote build tasks
