+++ package.json
@@ -23,24 +23,24 @@
     "typescript",
     "sass"
   ],
-  "main": "./electron/main.js",
+  "main": "./dist/electron/main.js",
   "private": true,
   "scripts": {
     "ng": "ng",
     "lint": "ng lint",
     "start": "webpack --watch",
     "start:web": "webpack-dev-server --content-base . --port 4200 --inline",
-    "build:electron:main": "tsc src/electron/main.ts --outDir dist/electron && copyfiles package.json dist && cd dist && npm install --prod && cd ..",
-    "build": "webpack --display-error-details && npm run build:electron:main",
+    "build:electron:main": "tsc src/electron/main.ts --outDir dist/electron",
+    "build": "npm run clean:dist && webpack --display-error-details && npm run build:electron:main",
     "build:prod": "cross-env NODE_ENV=production npm run build",
     "clean:dist": "rimraf ./dist",
-    "electron:serve": "npm run build:electron:main && electron ./dist --serve",
-    "electron:test": "electron ./dist",
-    "electron:dev": "npm run build && electron ./dist",
-    "electron:prod": "npm run build:prod && electron ./dist",
-    "electron:linux": "npm run build:prod && node package.js --asar --platform=linux --arch=x64",
-    "electron:windows": "npm run build:prod && node package.js",
-    "electron:mac": "npm run build:prod && node package.js --asar --platform=darwin --arch=x64",
+    "electron:serve": "npm run build:electron:main && electron . --serve",
+    "electron:test": "electron .",
+    "electron:dev": "npm run build && electron .",
+    "electron:prod": "npm run build:prod && electron .",
+    "electron:linux": "npm run build:prod && node build-assets/package.js --platform=linux",
+    "electron:windows": "npm run build:prod && node build-assets/package.js --platform=win",
+    "electron:mac": "npm run build:prod && node build-assets/package.js --platform=darwin",
     "test": "karma start ./karma.conf.js",
     "pree2e": "webdriver-manager update --standalone false --gecko false --quiet && npm run build",
     "e2e": "protractor ./protractor.conf.js"
